Make state and token_id cryptographically secure. (#22)
I have to mention that older versions of php may still return a cryptographically unsafe string (predictable one). For state is ok as it's hashed too, but token_id should be unpredictable. We should ensure that websites using this plugin use PHP 5.4.44 or bigger, 5.5.28 or bigger, or PHP 5.6.12 or bigger.